[refactor] enums 패키지 제거 후 domain 패키지에 통합
