+++ src/main/java/com/inq/wishhair/wesharewishhair/hairstyle/infra/query/HairStyleQueryRepository.java
@@ -2,8 +2,6 @@ package com.inq.wishhair.wesharewishhair.hairstyle.infra.query;
 
 import com.inq.wishhair.wesharewishhair.hairstyle.domain.HairStyle;
 import com.inq.wishhair.wesharewishhair.hairstyle.utils.HairRecommendCondition;
-import com.inq.wishhair.wesharewishhair.user.domain.FaceShape;
-import com.inq.wishhair.wesharewishhair.user.enums.Sex;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.domain.Slice;
