+++ src/main/java/com/inq/wishhair/wesharewishhair/hairstyle/utils/HairRecommendCondition.java
@@ -1,8 +1,8 @@
 package com.inq.wishhair.wesharewishhair.hairstyle.utils;
 
-import com.inq.wishhair.wesharewishhair.hairstyle.domain.hashtag.enums.Tag;
+import com.inq.wishhair.wesharewishhair.hairstyle.domain.hashtag.Tag;
 import com.inq.wishhair.wesharewishhair.user.domain.FaceShape;
-import com.inq.wishhair.wesharewishhair.user.enums.Sex;
+import com.inq.wishhair.wesharewishhair.user.domain.Sex;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
