+++ src/main/java/com/inq/wishhair/wesharewishhair/review/config/converter/ScoreConverter.java
@@ -2,7 +2,7 @@ package com.inq.wishhair.wesharewishhair.review.config.converter;
 
 import com.inq.wishhair.wesharewishhair.global.exception.ErrorCode;
 import com.inq.wishhair.wesharewishhair.global.exception.WishHairException;
-import com.inq.wishhair.wesharewishhair.review.domain.enums.Score;
+import com.inq.wishhair.wesharewishhair.review.domain.Score;
 import org.springframework.core.convert.converter.Converter;
 
 public class ScoreConverter implements Converter<String, Score> {
