+++ src/main/java/com/inq/wishhair/wesharewishhair/user/controller/dto/request/UserUpdateRequest.java
@@ -1,6 +1,6 @@
 package com.inq.wishhair.wesharewishhair.user.controller.dto.request;
 
-import com.inq.wishhair.wesharewishhair.user.enums.Sex;
+import com.inq.wishhair.wesharewishhair.user.domain.Sex;
 import lombok.AccessLevel;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
