+++ src/main/java/com/inq/wishhair/wesharewishhair/user/domain/Sex.java
@@ -1,0 +1,6 @@
+package com.inq.wishhair.wesharewishhair.user.domain;
+
+public enum Sex {
+    MAN,
+    WOMAN
+}
