--- src/main/java/com/inq/wishhair/wesharewishhair/user/enums/Sex.java
@@ -1,6 +1,0 @@
-package com.inq.wishhair.wesharewishhair.user.enums;
-
-public enum Sex {
-    MAN,
-    WOMAN
-}
