+++ src/main/java/com/inq/wishhair/wesharewishhair/user/service/dto/response/MyPageResponse.java
@@ -2,7 +2,7 @@ package com.inq.wishhair.wesharewishhair.user.service.dto.response;
 
 import com.inq.wishhair.wesharewishhair.review.service.dto.response.ReviewResponse;
 import com.inq.wishhair.wesharewishhair.user.domain.User;
-import com.inq.wishhair.wesharewishhair.user.enums.Sex;
+import com.inq.wishhair.wesharewishhair.user.domain.Sex;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
