+++ src/test/java/com/inq/wishhair/wesharewishhair/hairstyle/domain/hashtag/HashTagTest.java
@@ -2,7 +2,6 @@ package com.inq.wishhair.wesharewishhair.hairstyle.domain.hashtag;
 
 import com.inq.wishhair.wesharewishhair.global.fixture.HairStyleFixture;
 import com.inq.wishhair.wesharewishhair.hairstyle.domain.HairStyle;
-import com.inq.wishhair.wesharewishhair.hairstyle.domain.hashtag.enums.Tag;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
