+++ src/test/java/com/inq/wishhair/wesharewishhair/hairstyle/domain/hashtag/TagTest.java
@@ -1,6 +1,5 @@
 package com.inq.wishhair.wesharewishhair.hairstyle.domain.hashtag;
 
-import com.inq.wishhair.wesharewishhair.hairstyle.domain.hashtag.enums.Tag;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
