+++ src/test/java/com/inq/wishhair/wesharewishhair/user/utils/FlaskConnectorTest.java
@@ -3,7 +3,7 @@ package com.inq.wishhair.wesharewishhair.user.utils;
 import com.inq.wishhair.wesharewishhair.global.base.InfraTest;
 import com.inq.wishhair.wesharewishhair.global.exception.ErrorCode;
 import com.inq.wishhair.wesharewishhair.global.exception.WishHairException;
-import com.inq.wishhair.wesharewishhair.hairstyle.domain.hashtag.enums.Tag;
+import com.inq.wishhair.wesharewishhair.hairstyle.domain.hashtag.Tag;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
